requirement.txt
openenv-core
fastapi
uvicorn
requests
